Person class: add phone number to duplicate check
Persons are considered duplicates if they share the same name.

Different people sharing common names can be incorrectly flagged as
duplicates. A single non-unique (in the real world) identifier such as
name is insufficient to guarantee uniqueness as multiple distinct
individuals may share the same name.

Let's include a comparison of phone numbers, in addition to names, when
checking equality between two persons. Two persons are now treated as
the same if and only if both their names and phone numbers match.

Adding the phone number to the uniqueness check significantly reduces
false duplicate detections because it is much less likely for two
distinct people to share both a name and a phone number.